Automated QuickCode Functionality
QuickCodes must be specified when creating a new MenuCategory, MenuItem, MenuModificationCategory, or MenuModificationItem record. TRIGGERs automatically inserts the new QuickCode in said table with a NULL humanReadable description. 

Quick Codes are automatically removed from QuickCodes table, and any records referencing it in the MenuAssociations table are deleted. This occurs when any record from the above tables are deleted.