Automated QuickCode Functionality
QuickCodes must be specified for row insertion on any table that uses it as a primary key. QuickCodes and MenuAssocations are automatically manage via TRIGGERs to prevent putting the DB in an invalid state. Any newly defined quick code must have its humanReadable text adjusted manually in the QuickCodes table.