Added JS in cwpribble.js for moving items to ticketContainer
Each item that is clicked on in the menuContainer will be populated into the ticketContainer. The 'X' button does not work right now but that is an easy implementation I will add tomorrow and push once I have the PHP done. Load ticket had hard coded HTML in an echo statement. That has been commented out. There are also now four users, each with unique permission levels that are added via the third SQL script in an existing txt doc that was already populating menu items, quick codes, associations and mod categories/items.