Three Dumps For SQL. Completion Update coming 2/14
These dumps are the first three. There will be another dump but I need to go to bed so Ill do it tomorrow. That dump will be the modifications. The menu associations table dump will be updated as well once that is complete. This data is workable as it is and associated. The full modifications table will be uploaded tomorrow(2/14/2023) sometime in the afternoon. There is also the three files to create this data but I will push that tomorrow and in its own directory.