Adds checks to correctly enable/disable buttons
Buttons will now be disabled if they should not be pressed.

Fixed the seat dropdown to handle the maximum split of 0.

Added a php script to generate table shapes, tables, employees, and tickets.
This script also assigns tickets and employees to tables. It the php code that ran to help as a guide to team members writing sql statements.

Fixed duplicate ticket items displaying in ticket container.

Added a sql function to determine if all, some, or none of the splits associated with a ticket item have paid yet.