adds HostView wireframe to documentation
Adds capacity field to TableShapes table. Host needs to know how many people can fit in a specified table.

Updates createTicket function to include requestedTime parameter.
This invalidates any previous uses of createTicket.

Introduces timeReserved field to Tickets table. Allows new tickets to be created as a reservation or as-is as a walk in.