Create NewTicket.php
This is a PHP file that creates a new ticket for a restaurant waitlist system by processing a form that accepts a party nickname and size. The code executes a MySQL stored procedure to create the new ticket, retrieves the ticket number and time requested from the database, and displays them along with the party nickname and size. If there is an error, the code displays an error message. The file also includes an HTML form for entering the party information and a "Back" button to return to the waitlist page.