Implements WaitTimes.php
Wait Times are calculated by looking at a 5 to 60 minutes in the past. It takes the average of timeSeated - timeRequested. A main and max must be specified for party size.

This window refreshes automatically every 60 seconds, or if the time display is pressed/clicked. A timer at the bottom indicates the time remaining to refresh.... Adjusting the min/max/timeRange will invalidate wait time value and "Press to Refresh" is displayed.