All tableshapes now <path> drawn SVG's
Table shapes vectors (SVG) are now all created via <path> tags and use the draw attribute (d=Mxx xxx xx etc) to define the shape of each table. This change means that the width and height attributes are no longer usable within these table vectors. The way that they are defined as far as dimensions is now static. The scale (within the transform list) is now where the size is mutable. The dimensions, however, are not. This is intentional and prevents misdrawn table issues.