Comments added to config script
Comments added for the proverbial future developer that, in real production, would work on this source code in the future without any input from the original creators outside of the comments and documentation. This is a common issue with legacy codebases and the more comments, the better (to an extent: blocks should typically always be prefixed with a comment. Variable naming should be as indicative of the contents as possible, but not discursive). The introduction of a new convention for comments/statements that are left in a state that is not intended to be inveterate on production. Within a comment meant to indicate a stopgap, the exact string, (to be removed), should be added so all occurrences of temporary solutions/comments can be found and removed/changed via Find. In VS Code, 'CTRL+Shift+F 'on the CentRes directory will find these instances, making the digging for necessary removals much easier before the final deliverable is integrated and polished.