Older and Newer Version of the layout. FEEDBACK
Please provide feed back on the two. The layout_svg.png is an image of the layout with a lot of tables and is probably a more difficult implementation as on most devices I have tested, It wouldn't be possible to see text within tables. The second one, called layout_svg_basic.png is an image of another version ofa more basic and easily manipulable layout. The second one (obviously) uses less structure shapes (there is a table in the db dedicated to these and they will be queried and placed just like tables but will not be updatable by the HOST. Feedback is what I am looking for here.