CentResPopulateTableshapesAndTables.txt update/add
The file called CentResPopulate_tableshapes is updated and now has been renamed to the name in the description of this commit. The reference to this file in the resetDB.php file, located in the same dir (Configuration) has been updated to reflect this change, getting the file with it's updated name. The table called 'tables' will be within this script as well. That is coming after.