Fixes numerous bugs in ServerView
Fixes 404 Error for displayInterface (Failed to load Resource)

Double-Clicking on a table in HostView as manager once again takes you to that table/ticket in HostView.

Fixes bug where clicking a menu item in ServerView fails to add it to selected Ticket/Seat/Split.

Fixes selected ticket item styling bug: If multiple items are moved to another seat/split, the first item would only have the "selected" class applied to it, and not "multiselect".

This created an additional bug that manifested during the Sprint 2 demo dry run-through. Deleting an item from the ticket after moving multiple items to a different split/set would result in an invalid state and cause the removal operation to fail.