Update CentResDBFunctionsAndMethods.txt
I've modified ticketItemPrice Function to calculate the number of splits by counting the number of Ticket Items associated with the same ticket Id as the given ticketItem Id and the same split Flag(splitFlg). If there is more than one split, the price is divided by the number of splits.