Numerous bug fixes in ServerView.php and ticket.php
fixes database splitString function to use a loop.
Split string now displays in ServerView in ascending order.

fixes ServerView "DELIVER" button to match "SUBMIT" and "CANCEL" buttons...
It will be enabled if there are any ticket items ready to be delivered.
Clicking this button will mark all "Ready" ticket items as delivered.

Fixes bug in ticket.php. $_POST['ticketItem'] was not being unset.
This caused a ticket item to fail to remove if it was just added.

Changes "Seat" dropdown to reflect number of people in a ticket partySize.
If for example there are 5 people, there will be 5 seats available.