Implements guards against malicious DB resets
Configuration folder has been moved to CodeRepo.

Access to any of those php files in that folder require manager access or an error with connecting to the database. If you are not logged in as a manager, you will be redirected to the login page.

Scripts have removed creating a manager.

If no manager is detected when login page is accessed, you are redirected to CreateUserView.php. Here you will create a manager with maximum permission levels.

Attempting to access CreateUserView.php with an existing manager in the system will redirect you to the login page.