Adjusts database to assign groupIndex to ticket items when submitted
calling submitPendingTicketItems now groups the submitted ticket items into the same group. groupIndex is an int.

groupId is a decimal number.
The whole number represents the ticketId/number.
The decimal portion represents the group number within that ticjet number.
This convention allows for up to 99 submissions aka pressing the SUBMIT button on the server view before reaching an invalid state.