Introduces transient logins.
Login screen will now display a list of routes after an employee selects "Terminal Access".
When the employee selects a route, they are logged in with a temporary username. USERNAME~ROUTE and redirected to BackOfHouseTest.php

sessionHeader logs the temporary account out and deletes it.

Any window with a transient login has an orange "Log In" button at the top right of sessionHeader. This takes the place of where the "Logout" button is for regular logins. Pressing "Log In" takes takes you back to the login screen.

sessionInfo listener is not running for transient logins.