Terminals now have the ability to filter traffic.
Building upon Donovan's adjustments to the database, Back of House now will ignore any ticket group traffic that does not match the terminal's route.

Fixes previous push that caused the DB reset script to fail.

Adjusts formula used to auto-generate a quick code in each of the 4 menu related tables. It is now impossible for auto-generation to compuse a quick code already in use.