Fixes duplicate key entry error when updateing menu item
Also redesigns ticket group frame to have a fixed "descriptor" header. If the item list overflows, you can now scroll underneath the header.

To accommodate this tweak, the header will now flash black for a second to indicate a change. The shrinking visual cue looked jarring and had to be replaced.