Fixes error in modOptionUtility.js generateModOptionDiv function
The function was incorrectly omitting the price for a single checkbox/radio when a single value was provided in quantifierString. The price would fail to display in ticket.php and not correctly apply to the ticket's calculatedPriceWithMods field.

Refactored the code for final release. Resources/php was renamed to Resources/PHP to match the folder naming convention in Resources. connect_disconnect.php was renamed to dbConnection.php