Fixes bug introduced in last commit by removing svgManipulation.
Also corrects a long-time visual bug in the ticket for Server View.
If you tried to add a menu item but don't have a seat or split selected, a flashing red alert at the top of the ticket should appear.

If you were scrolled too far down, the alert would not be visible, and not provide visible feedback. This causes the Server View appear unresponsive when adding a menu item.

The solution was just scroll to the top and show the red flashing error message.

Finally, corrects closeBusinessDay db function. After the data is dumped, the businessDay value in the Config table was never updated. This let to the businessDay being stuck on the day the database was installed.