Numerous across the board bug fixes and improvements
CloseBusinessDay now logs out any active employees.

Stylized alert in server view now appears when ticket is removed from table and manager is viewing it. They are returned to host view after clicking ok.

Manager now has the ability to remove ticket items at any time when looking at a ticket in server view.

Terminal View now displays items removed by manager with the grey color and strikethrough text.

If a server removes an item after being submitted, it will take a manager to officially remove it. The cost is still being applied to the ticket/split at this time.

After a manager removes the item, it will disappear from the server view and not be counted towards the subtotal.

Removed ticket items that are now invisible on the Server View will still persist on Terminal View until the close button is pressed.