expands menu object title hiding functionality
If any word in a menu object title starts with a period [ . ], it will be hidden in the menu, ticket, and terminal views.

This functionality was introduced to allow to have items that appear exactly the same, but can take on different properties.

A usage example of this might be having a "soup of the day that rotates daily" Suffixing the title with .mon, .tue, .wed..... etc could facilitate this requirement.