One change, One prompt - simple
Changed mod window to display mod - action in a column format using a flex definition.
Added prompt with sample refresher and helpful link (will be removed) as SQL comment. This is for adding a check in the end of business day's current business day value in table-`Config`. If it is not directly the subsequent day, the bDay (business day) value will change to the correct 'current' date to avoid inaccurate data logging.